requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
celery
redis
mlflow
transformers
torch
pydantic
pydantic-settings
python-dotenv
aiofiles
scikit-learn>=1.0.2
pandas>=1.5.0
numpy>=1.23.0
requests>=2.31.0
nltk==3.8.1 